Move staff admin components to /admin directory
