+++ app/javascript/packs/application/App.js
@@ -4,13 +4,13 @@ import { connect } from "react-redux";
 import PropTypes from "prop-types";
 
 import Navigation from "./components/Navigation";
-import LoginForm from "./components/LoginForm";
-import CreateAccountForm from "./components/CreateAccountForm";
-import StaffUserDashboard from "./components/StaffUserDashboard";
-import CreateTourForm from "./components/CreateTourForm";
-import UpdateTourForm from "./components/UpdateTourForm";
-import TourContainer from "./components/TourContainer";
-import UpdatePointForm from "./components/UpdatePointForm";
+import LoginForm from "./components/admin/LoginForm";
+import CreateAccountForm from "./components/admin/CreateAccountForm";
+import StaffUserDashboard from "./components/admin/StaffUserDashboard";
+import CreateTourForm from "./components/admin/CreateTourForm";
+import UpdateTourForm from "./components/admin/UpdateTourForm";
+import TourContainer from "./components/admin/TourContainer";
+import UpdatePointForm from "./components/admin/UpdatePointForm";
 import api from "./api";
 import * as staffUserActions from "./redux/staffUser/actions";
 import * as staffUserSelectors from "./redux/staffUser/selectors";
