+++ app/javascript/packs/application/components/admin/ImageInput.js
