+++ app/javascript/packs/application/components/admin/Point.test.js
