+++ app/javascript/packs/application/components/admin/PointFormFields.js
