+++ app/javascript/packs/application/components/admin/PointFormFields.test.js
