+++ app/javascript/packs/application/components/admin/TourFormFields.js
